Support for GHC reading DWARF
Turned out to be more complicated than I thought (doesn't it always). To
do this properly, we have to keep track of a lot of things like where
exactly shared libraries are mapped to. In the end, I'm not even sure it
is worth the effort I put into it, given that the most interesting chunk
of code - the program - often is relocated quite trivially.

Also note that this is, again, *very* unportable as both Windows and Mac
Os have different binary formats. So this dance might have to be repeated
a few times in future.