Added file name to source debug annotation
The reason that I didn't add it before was worries over file size. But
that's actually pretty silly - on one hand we have a lot less granularity
for source code annotations now and on the other core ticks are actually
blowing up the event log quite a bit more already.