Use libelf to read out debug data section instead of hs_hpc_module
This change was prompted by "this_mod" obviously not always being
defined in the backend - which means that the debug code can't assume
that the module name is available. I removed the "Module" parameter for
a number of functions as well to make it harder to fall into that trap.

This is arguably a better way of doing it: This allows external programs
such as strip to safely discard the info after the program was built.

(untested for now, as I can't bring GHC to build due to probably
unrelated issues)