LLVM codegen monad
This is arguably more elegant than passing LlvmEnv as well as a bunch
of other things around manually. Most importantly, it makes it easy to
add new features, like the more flexible solution for declaring
variables as used.

On the other hand, this requires *major* refactoring. Not sure it's
worth the trouble come merge time.