Declare debug data as local, but used
I was originally pushing into making the whole debug data blob "appending",
but it seems this in fact doesn't work unless you use the LLVM linker
(which we won't anytime soon). But this works as well, as the RTS code
doesn't really depend on the debug sections being in one piece.