Attempt to repair the "text/str" RULE
Idea is to automatically skip the conversion to String where possible -
Pretty can work with the original C-Strings just as well. There is a
rule for that already, but it doesn't fire due to string fusion getting
in the way.

This is an attempt to restore the original behaviour. Unluckily, it comes
with worse performance in corner cases, see comments.