Set !llvm.dbg.sp in LLVM code
Specification says we ought to do that. This might have lifted the need
for adding !dbg annotations to the source as well.

Note that the format for metadata node collections is pretty inconsistent
with the way metadata normally look like. Even the docs seem to get
confused about it. On our side we have to thread a different code path all
the way through just so we don't generate a "metadata" too much. Ugly.