Experimental: Generate case alternatives into fresh procs
This has the advantage that we can associate DWARF with them. The trick
here is that at least with LLVM and -O2, this transformation will be
reversed by "opt". However LLVM properly documents this process in DWARF,
marking each alternative as an "inlined subroutine". That's enough data
for us to tease apart the costs.