Moved debug data generation to separate module, write using Binary
This makes a lot more sense than the old design - not only is debug
data generation something that isn't really specific to the LLVM backend,
but it is also pretty awkward and slow to construct everything as
pseudo-structures, when we are pretty sure about the exact binary
layout we want to have in the end (namely, the event log format).

The new code does everything using fast in-memory buffers, using a slight
extension of BinHandle. This should finally improve compilation speeds a
bit.