Move tick context lookups to code generation
By looping in the finished map and looking up the context lazily. Why be
awkward when you can be evil instead?

Seriously, this is really hard to get around without restructuring code
generation so it decides earlier what label to assign to procedures. This
makes it at least straightforward to work with the product.