Reworked debug in code generation
The new idea is to exploit that codegen has the CgStmt wrapper for the
actual Cmm code in order to easily emit new blocks "inline". We can
use this infrastructure by inserting debug and context annotations about
the generated code.

When the code is then actually printed, we collect together all this
information and generate the tick map. Note that context (= timParent)
info can now originate from a different place then the debug
annotations (= timTicks), so we must make sure to merge suitably. Sadly,
the whole process still requires the tick map to get looped in...

I'm still a lot more pleased with this code, mainly because it is far less
intrusive to the codegen architecture. In fact, a lot of this commit is a
clean rollback to master!

As an added bonus, this change allows for far finer granularity (block
level!). Now we just need to figure out how to cleanly map this to Core
and we can finally talk about profiling "case" alternatives against each
other, amongst other things.