Fixed assert in code generation
We want context nodes to get embedded into the parent context in all cases.
Yet this is not true for all CG statements, as there are contexts where
certain FCode actions are not supposed to emit any statements. Most
notably, some only emits out-of-line procedures - which unluckily is the
exact position where context annotations are emitted.

So we hack around it a bit - stateIncUsageEval already does the right
thing despite its name, we just need to teach GHC that it's okay to see
context entries at this point.