Support distinguishing case alternatives, output Core in binary
The binary Core thing might seem a bit dubious - the reasoning is mainly
that I want the flexibility of freely organizing the Core code in the
profiler, and I'd have a pretty rough time with that if I had to squeeze
everything through the existing pretty-printer - like the constructor
annotations introduced here. On top of that, I actually care about size,
GHC profiles weight 350M already, which is mainly Core. So a partly binary
representation seemed like a straightforward choice.

We'll see whether this particular design decision stands the test of time.