Increased source annotation density
The main idea is that we end up with source code annotations on *all*
interesting code blocks, so we can track more closely how exactly code
gets "fused" together.

Note though that this means a *lot* of source note duplication. To counter
this, this patch also starts eliminating source notes that are "included"
in other source note (thinking about sub-expressions here, mainly). That
seems to reduce the number of source code annotations to a manageable
amount.