Don't treat lax tickishs as lambdas for purpose of occurrence analysis
We actually want stuff to get inlined into it. I also feel like this
should have some sort of special case for when the inline site actually
belongs to the same SCC / tickish as the one getting inlined. The
simplifier seems to pick stuff out of expressions fairly often, and I
think the code as it is will prevent them from getting inlined again...