Profile life heap objects
This uses the heap census from cost centre profiling to produce
code pointers for our profiling framework. This is a pretty
cheap and easy change, and is enough to deduce pretty much
everything -hm, -hd and -hy know without requiring PROFILING.
In fact, we can know identify the closures taking up most space
on the heap by their Core and source code, which could turn
out to be pretty useful.