Declare LLVM aliases of external entities as external
Normally all aliases get resolved by opt, but for large object files LLVM
doesn't do that and leaves some aliases in the assembler file. This leads
to the assembler regarding the symbol as local, and dynamic linking
failing at some point down the line.

This fixes it by declaring the aliases external. Surprisingly, this does
not cause them to become part of the symbol table - I'm not sure I
understand the logic here, but let's go with it for the moment.