"Annotated" pretty-printing style
This allows inserting zero-width annotations to the output that can be
used later to deduce what part of (here) the Core tree ended up where in
the pretty-printed dump.

This is obviously somewhat hackish - I would prefer to have a
way to teach HughesPJ to give me back line number information about what
it printed. But I haven't been able to think of a good way of doing that
yet... So for now let's do it this way.