Replace TickMap by new CmmTicks solution
A lot of reverts to pre-tick-map state here. We use the occasion for
some more refactoring in the LLVM backend: It probably makes more sense
for LlvmCodeGen.CodeGen to drive LlvmMeta for procedure and block meta
data.

The big missing piece is that we have no debug data generation for TS
whatsoever at this point - just the meta data (and therefore DWARF).