DWARF generation for the native code backend
There's two parts to this - generating .dwarf_abbrev / .dwarf_info takes
are of communicating blocks and procedure structure to the compiler, and
will act as the main source of information for the RTS. We generate the
code for this directly, as it is not a very complicated format.

On the other hand, for proper debuggers we want direct source code links,
which is encoded in .debug_loc. Here we actually delegate the generation
to the assembler using using the ".file" and ".loc" directives. This
means we don't have to care about the rather incrinate finicky state-machine
compression algorithm DWARF specifies. LLVM takes the same route, so
we don't have to feel too bad about it.

On the other hand, this means that we need to interleave debugging info
with assembly code, so we have to meddle with platform-specific code here.
At this point we only support X86, but it *should* be easy to port.
Furthermore, tracking the ".file" annotations unfortunately requires a
significant amount of plumbing... A monad could really do wonders here.