Allow Cmm code to define unwind rules
This is a bit more general than what we had before. Basically now we
can write Cmm code like

   unwind Sp = Sp + 8;

in order to tell the DWARF generation to set the unwind rule to "increase
Sp by 8" at this point. The syntax is pretty general and could probably
be extended at some point - right now back-end support is however limited
to
1) exactly "unwind Sp = Sp + x" expressions and
2) them exactly appearing as the first statement in a block

This is however enough to provide proper unwinding instructions for a
number of Cmm routines that use an explicit stack layout. Maybe at some
point we could even get the debugger to make the transition to C land
this way?