Improve strictness when generating debug data with LLVM
As we have to emit all debug data as one block at the end of LLVM code
generation, we must be careful that we don't keep all Cmm data alive
until that point as well. This makes memory consumption on large code
files a bit less bad, but though there's probably still room for
improvement.