Generate debug information per Cmm group instead of proc
This allows Cmm blocks to set themselves as context of blocks in other
procs, as long as they get generated in the same group. This solution
is a bit awkward, as I don't feel there is any strong reason for block
labels to stay stable. We will probably only do it for entry points, but
still...