Emit ticks before heap check
This makes sure that we get the ticks in entry blocks, and properly cover
the cost of the heap check. This is generally the right thing to do as
long as the tick in question doesn't track entry counts: Those would be
wrong due to re-execution of the entry block after a failed heap check.