Revert eagerly taking over C/LD/CXXFLAGS
Seems I was to blame for the spurious "-g" parameters getting passed to
the assembler after all. I had introduced this in order to be able
to link against libraries (say, libdwarf) in non-standard locations. But
that's probably not a very common problem, and should have a more elegant
solution (--with-dwarf-libraries?) anyway.