[untested] Add unwind information for SAVE_THREAD_STATE
This is an example of how to set more complicated unwind rules - here
for returnToSched & co. Note that we have to work around a few issues here:
The unwind declaration needs to be the first node in the block, so we
move SAVE_THREAD_STATE accordingly. That's ugly - a system that handles
unwind rules in the middle of a block would be better.