Tick scopes
This patch solves the scoping problem of CmmTick nodes: If we just put
CmmTicks into blocks we have no idea what exactly they are meant to cover.
Here we introduce nested scopes, represented as lists of uniques. The
"nesting" relation is given by the subset relation. For example a tick
declared in a block with, say, scope [b,a] now scopes over all blocks
that have at least a tick scope of [b,a], so for example also [c,b,a].

Notes:

* This makes it easy to express most optimisations: It is both easy to
  generate new blocks that share all ticks with existing blocks. It is
  especially possible to merge blocks to have combined contexts, simply
  by merging the scope lists. If this happens, we actually end up with
  an (acyclic) scope graph instead.

* Given that the code often passes Cmm around "head-less", we have to
  make sure that its intended scope does not get lost. To keep the amount
  of passing-around to a minimum we define a CmmAGraphScoped type synonym
  here that just bundles the scope with a portion of Cmm to be assembled
  later.

* We introduce new scopes at somewhat random places, aligning with
  getCode calls. This works surprisingly well, but we might have to
  add new scopes into the mix later on if we find things too be too
  coarse-grained.