Heap allocation profiling
This is one of the cheapest possible ways to get profiling data: The
nursery grows block-wise, with each block getting requested separately
after the last block filled up. We simply fill an array with the sources
of these requests, and get a nice overview of allocation hot spots out
of it.