Stack allocation sampling
This is a bit of an experiment - we can theoretically re-use the heap
allocation profiling facilities for identification of instances where
we allocate a lot of stack space (enough to warrant requesting new blocks!).
This is however less useful due to the fact that the allocation of new
stack blocks is, in fact, quite rare. We might have to decrease stack
chunk size (-kc?) to get anything useful out of it.

Also the implementation is pretty hacky...